refactor reach spell: methods delegated out of message into inventory and spell as needed
