+++ src/main/java/com/lothrazar/cyclicmagic/ModMain.java
@@ -51,7 +51,7 @@ public class ModMain {
 		network.registerMessage(MessageToggleSpell.class, MessageToggleSpell.class, packetID++, Side.SERVER);
 		network.registerMessage(MessageParticle.class, MessageParticle.class, packetID++, Side.CLIENT);
 		network.registerMessage(MessageOpenSpellbook.class, MessageOpenSpellbook.class, packetID++, Side.CLIENT);
-		network.registerMessage(MessagePlaceBlock.class, MessagePlaceBlock.class, packetID++, Side.SERVER);
+		network.registerMessage(MessageSpellReach.class, MessageSpellReach.class, packetID++, Side.SERVER);
 		network.registerMessage(MessageToggleBuild.class, MessageToggleBuild.class, packetID++, Side.SERVER);
